Rename points.tsx again.
