+++ src/components/plot/points.tsx
